% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.jointCI.R
\name{my.jointCI}
\alias{my.jointCI}
\title{my.jointCI}
\usage{
my.jointCI(data, type = c("T2", "bonferroni"), alpha)
}
\arguments{
\item{data}{multivariate data}

\item{type}{If type="T2",calculate the T2 joint CI.If type="bonferroni",calculate bonferroni joint CI.}

\item{alpha}{level}
}
\value{
the joint confidence interval of multivariate data
}
\description{
give the joint confidence interval of multivariate data
}
\examples{
\dontrun{
x <- rnorm(100,10,4)
y <- rnorm(100,20,8)
data <- matrix(data = c(x,y),100,2)
my.jointCI(data,type = "T2",alpha=0.05)
}
}
