% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.ppca.R
\name{my.ppca}
\alias{my.ppca}
\title{my.ppca}
\usage{
my.ppca(q, S)
}
\arguments{
\item{q}{numbers of PC}

\item{S}{the covariance matrix}
}
\value{
the maximum-likelihood estimator for variance and the weight matrix
}
\description{
give the covariance matrix,calculate the probabilistic principal component by MLE.
}
\examples{
\dontrun{
library(elasticnet) 
data(pitprops)
my.ppca(q=3,S=pitprops)
}
}
