% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbnC}
\alias{rbnC}
\title{A Gibbs sampler to generate a bivariate normal chain using Rcpp}
\usage{
rbnC(N, mu1, mu2, sigma1, sigma2, rho)
}
\arguments{
\item{N}{the length of chain}

\item{mu1}{Initial parameters}

\item{mu2}{Initial parameters}

\item{sigma1}{Initial parameters}

\item{sigma2}{Initial parameters}

\item{rho}{Initial parameters}
}
\value{
a a Gibbs sample chain
}
\description{
A Gibbs sampler to generate a bivariate normal chain using Rcpp
}
\examples{
\dontrun{
rbnC(1000,0,0,1,1,0.9)
}
}
