\name{StatComp22038-package}
\alias{StatComp22038-package}
\alias{StatComp22038}
\docType{package}
\title{
  Final homework for 'Statistical Computing' course
}
\description{
  R version of SPCA and PPCA.Rcpp version of binoraml chain.
}
\author{
Jiaying
Maintainer: Jiaying <jy100@mai.ustc.edu.cn>
}
\keyword{ package }

